pyproject.toml
[project]
name = "h1b"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]
